- Add MP_Pack::PackZip - Add WIC_PROTOCOL_VER enumeration - Add MESSAING_CLAN_CREATE_REQUEST basic test - Modify project to work with x64 better

+++ MLiveAccountServer/src/Server Modules/JSON.h
@@ -122,7 +122,7 @@ public:
 			return false;
 
 		if(buffer)
-			MC_Misc::MultibyteToUnicode(json_string_value(obj), buffer, bufferSize / sizeof(wchar_t));
+			MC_Misc::MultibyteToUnicode(json_string_value(obj), buffer, (uint)(bufferSize / sizeof(wchar_t)));
 
 		return true;
 	}
